[Issue 38] move fractal dispatch from worker to EscapeTime
Worker is part of an engine.  Dispatch is fundamental to the EscapeTime
module, not any engine.
